+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # tidepool-platform-client
 Client-side library to interact with the [Tidepool](http://tidepool.org/) platform.
 
-## [0.1.0] 
+## 0.2.0 - 2020-04-24
+###  Added
+- PT-1251 Add TIR metric in search result page
+
+## [0.1.0]
 ###  Added
 - Integrate latest change from [Tidepool 0.42.0](https://github.com/tidepool-org/platform-client/releases/tag/v0.42.0)